+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ new version (on deck)
   tidyr's ``pivot_longer`` function. Also added an example notebook. @samukweku
 - [ENH] Fixed code to returns error if `fill_value` is not a dictionary. @samukweku
 - [INF] Welcome bot (.github/config.yml) for new users added. Issue #739. @samukweku
+- [ENH] ``pivot_longer`` can handle multiple values in paired columns, and can reshape
+  using a list/tuple of regular expressions in ``names_pattern``. @samukweku
 
 v0.20.9
 =======